review comments: also merged the divided classes
