+++ modules/swift/examples/code_snippets/SampleCodeTest.swift
@@ -24,8 +24,8 @@ import CoreML
 class SampleCodeTest {
     /**
      For consistency in code snippets:
-        1. we will use `self.database`/ `database` for database, query, replicator-db related code snippets.
-        2. we will use `self.otherDB` / `otherDB` for listener-db
+     1. we will use `self.database`/ `database` for database, query, replicator-db related code snippets.
+     2. we will use `self.otherDB` / `otherDB` for listener-db
      */
     
     var database: Database!
@@ -33,23 +33,2430 @@ class SampleCodeTest {
     
     /**
      For consistency:
-        1. we will use replicator with `self.replicator` and listener with `self.listener`
+     1. we will use replicator with `self.replicator` and listener with `self.listener`
      */
     var replicator: Replicator!
     var listener: URLEndpointListener!
+    
+    var replicatorsToPeers = [String: Replicator]()
+    var replicatorListenerTokens = [String: Any]()
+    
+    // MARK: Database
+    
+    func dontTestNewDatabase() throws {
+        let userDb: Database = self.database;
+        // tag::new-database[]
+        var database: Database!
+        do {
+            database = try Database(name: "my-database")
+        } catch {
+            print(error)
+        }
+        // end::new-database[]
+        
+        // tag::close-database[]
+        do {
+            try userDb.close()
+        }
+        // end::close-database[]
+        print(database)
+    }
+    
+#if COUCHBASE_ENTERPRISE
+    func dontTestDatabaseEncryption() throws {
+        // tag::database-encryption[]
+        let config = DatabaseConfiguration()
+        config.encryptionKey = EncryptionKey.password("secretpassword")
+        self.database = try Database(name: "my-database", config: config)
+        // end::database-encryption[]
+    }
+#endif
+    
+    func dontTestLogging() throws {
+        // tag::logging[]
+        // verbose / replicator
+        Database.log.console.level = .verbose
+        Database.log.console.domains = .replicator
+        
+        // verbose / query
+        Database.log.console.level = .verbose
+        Database.log.console.domains = .query
+        // end::logging[]
+    }
+    
+    func dontTestConsoleLogging() throws {
+        // tag::console-logging[]
+        Database.log.console.domains = .all // <.>
+        Database.log.console.level = .verbose // <.>
+        
+        // end::console-logging[]
+        // tag::console-logging-db[]
+        
+        Database.log.console.domains = .database
+        
+        // end::console-logging-db[]
+    }
+    
+    func dontTestFileLogging() throws {
+        // tag::file-logging[]
+        let tempFolder = NSTemporaryDirectory().appending("cbllog")
+        let config = LogFileConfiguration(directory: tempFolder) // <.>
+        config.usePlainText = true // <.>
+        config.maxSize = 1024 // <.>
+        Database.log.file.config = config // <.>
+        Database.log.file.level = .info // <.>
+        // end::file-logging[]
+    }
+    
+    func dontTestEnableCustomLogging() throws {
+        // tag::set-custom-logging[]
+        let logger = LogTestLogger(.warning)
+        Database.log.custom =  logger // <.>
+        // end::set-custom-logging[]
+    }
+    
+    func dontTestLoadingPrebuilt() throws {
+        // tag::prebuilt-database[]
+        // Note: Getting the path to a database is platform-specific.
+        // For iOS you need to get the path from the main bundle.
+        let path = Bundle.main.path(forResource: "travel-sample", ofType: "cblite2")!
+        if !Database.exists(withName: "travel-sample") {
+            do {
+                try Database.copy(fromPath: path, toDatabase: "travel-sample", withConfig: nil)
+            } catch {
+                fatalError("Could not load pre-built database")
+            }
+        }
+        // end::prebuilt-database[]
+    }
+    
+    // MARK: Document
+    
+    func dontTestInitializer() throws {
+        // tag::initializer[]
+        let newTask = MutableDocument()
+            .setString("task", forKey: "type")
+            .setString("todo", forKey: "owner")
+            .setDate(Date(), forKey: "createdAt")
+        try database.saveDocument(newTask)
+        // end::initializer[]
+    }
+    
+    func dontTestMutability() throws {
+        // tag::update-document[]
+        guard let document = database.document(withID: "xyz") else { return }
+        let mutableDocument = document.toMutable()
+        mutableDocument.setString("apples", forKey: "name")
+        try database.saveDocument(mutableDocument)
+        // end::update-document[]
+    }
+    
+    func dontTestTypedAcessors() throws {
+        let newTask = MutableDocument()
+        
+        // tag::date-getter[]
+        newTask.setValue(Date(), forKey: "createdAt")
+        let date = newTask.date(forKey: "createdAt")
+        // end::date-getter[]
+        
+        // tag::to-dictionary[]
+        print(newTask.toDictionary())  // <.>
+        
+        // end::to-dictionary[]
+        
+        // tag::to-json[]
+        print(newTask.toJSON()) // <.>
+        
+        // end::to-json[]
+        
+        print("\(date!)")
+    }
+    
+    func dontTestBatchOperations() throws {
+        // tag::batch[]
+        do {
+            try database.inBatch {
+                for i in 0...10 {
+                    let doc = MutableDocument()
+                    doc.setValue("user", forKey: "type")
+                    doc.setValue("user \(i)", forKey: "name")
+                    doc.setBoolean(false, forKey: "admin")
+                    try database.saveDocument(doc)
+                    print("saved user document \(doc.string(forKey: "name")!)")
+                }
+            }
+        } catch let error {
+            print(error.localizedDescription)
+        }
+        // end::batch[]
+    }
+    
+    func dontTestChangeListener() throws {
+        // tag::document-listener[]
+        database.addDocumentChangeListener(withID: "user.john") { (change) in
+            if let document = self.database.document(withID: change.documentID) {
+                print("Status :: \(document.string(forKey: "verified_account")!)")
+            }
+        }
+        // end::document-listener[]
+    }
+    
+    func dontTestDocumentExpiration() throws {
+        // tag::document-expiration[]
+        // Purge the document one day from now
+        let ttl = Calendar.current.date(byAdding: .day, value: 1, to: Date())
+        try database.setDocumentExpiration(withID: "doc123", expiration: ttl)
+        
+        // Reset expiration
+        try database.setDocumentExpiration(withID: "doc1", expiration: nil)
+        
+        // Query documents that will be expired in less than five minutes
+        let fiveMinutesFromNow = Date(timeIntervalSinceNow: 60 * 5).timeIntervalSince1970
+        let query = QueryBuilder
+            .select(SelectResult.expression(Meta.id))
+            .from(DataSource.database(database))
+            .where(
+                Meta.expiration.lessThan(
+                    Expression.double(fiveMinutesFromNow)
+                )
+            )
+        // end::document-expiration[]
+        print(query)
+    }
+    
+    func dontTestBlob() throws {
+#if TARGET_OS_IPHONE
+        let newTask = MutableDocument()
+        var image: UIImage!
+        
+        // tag::blob[]
+        let appleImage = UIImage(named: "avatar.jpg")!
+        let imageData = UIImageJPEGRepresentation(appleImage, 1)! // <.>
+        
+        let blob = Blob(contentType: "image/jpeg", data: imageData) // <.>
+        newTask.setBlob(blob, forKey: "avatar") // <.>
+        try database.saveDocument(newTask)
+        
+        if let taskBlob = newTask.blob(forKey: "image") {
+            image = UIImage(data: taskBlob.content!)
+        }
+        // end::blob[]
+        
+        print("\(image)")
+#endif
+    }
+    
+    // MARK: Query
+    
+    func dontTestQueryGetAll() throws {
+        // tag::query-get-all[]
+        let database = try Database(name: "hotel")
+        let query = QueryBuilder
+            .select(SelectResult.expression(Meta.id).as("metaId"))
+            .from(DataSource.database(database))
+        
+        // end::query-get-all[]
+        print(query)
+    }
+    
+    func dontTestIndexing() throws {
+        // tag::query-index[]
+        let index = IndexBuilder.valueIndex(items: ValueIndexItem.expression(Expression.property("type")),
+                                            ValueIndexItem.expression(Expression.property("name")))
+        try database.createIndex(index, withName: "TypeNameIndex")
+        // end::query-index[]
+    }
+    
+    func dontTestSelectMeta() throws {
+        // tag::query-select-meta[]
+        let query = QueryBuilder
+            .select(SelectResult.expression(Meta.id))
+            .from(DataSource.database(database))
+        
+        do {
+            for result in try query.execute() {
+                print("document id :: \(result.string(forKey: "id")!)")
+                print("document name :: \(result.string(forKey: "name")!)")
+            }
+        } catch {
+            print(error)
+        }
+        // end::query-select-meta[]
+    }
+    
+    
+    func dontTestSelectProps() throws {
+        // tag::query-select-props[]
+        let query = QueryBuilder
+            .select(
+                SelectResult.expression(Meta.id),
+                SelectResult.property("type"),
+                SelectResult.property("name")
+            )
+            .from(DataSource.database(database))
+        
+        do {
+            for result in try query.execute() {
+                print("document id :: \(result.string(forKey: "id")!)")
+                print("document name :: \(result.string(forKey: "name")!)")
+            }
+        } catch {
+            print(error)
+        }
+        // end::query-select-props[]
+    }
+    
+    func dontTestSelectAll() throws {
+        var query: Query
+        
+        // tag::query-select-all[]
+        query = QueryBuilder
+            .select(SelectResult.all())
+            .from(DataSource.database(database))
+        // end::query-select-all[]
+        
+        // tag::live-query[]
+        query = QueryBuilder
+            .select(SelectResult.all())
+            .from(DataSource.database(database)) // <.>
+        
+        // Adds a query change listener.
+        // Changes will be posted on the main queue.
+        let token = query.addChangeListener { (change) in // <.>
+            for result in change.results! {
+                print(result.keys)
+                /* Update UI */
+            }
+        } // <.>
+        
+        // end::live-query[]
+        
+        // tag::stop-live-query[]
+        query.removeChangeListener(withToken: token) // <.>
+        
+        // end::stop-live-query[]
+        
+        print("\(query)")
+    }
+    
+    func dontTestWhere() throws {
+        // tag::query-where[]
+        let query = QueryBuilder
+            .select(SelectResult.all())
+            .from(DataSource.database(database))
+            .where(Expression.property("type").equalTo(Expression.string("hotel")))
+            .limit(Expression.int(10))
+        
+        do {
+            for result in try query.execute() {
+                if let dict = result.dictionary(forKey: "travel-sample") {
+                    print("document name :: \(dict.string(forKey: "name")!)")
+                }
+            }
+        } catch {
+            print(error)
+        }
+        // end::query-where[]
+    }
+    
+    func dontTestQueryDeletedDocuments() throws {
+        // tag::query-deleted-documents[]
+        // Query documents that have been deleted
+        let query = QueryBuilder
+            .select(SelectResult.expression(Meta.id))
+            .from(DataSource.database(database))
+            .where(Meta.isDeleted)
+        // end::query-deleted-documents[]
+        print(query)
+    }
+    
+    func dontTestCollectionOperatorContains() throws {
+        // tag::query-collection-operator-contains[]
+        let query = QueryBuilder
+            .select(
+                SelectResult.expression(Meta.id),
+                SelectResult.property("name"),
+                SelectResult.property("public_likes")
+            )
+            .from(DataSource.database(database))
+            .where(Expression.property("type").equalTo(Expression.string("hotel"))
+                    .and(ArrayFunction.contains(Expression.property("public_likes"),
+                                                value: Expression.string("Armani Langworth")))
+            )
+        
+        do {
+            for result in try query.execute() {
+                print("public_likes :: \(result.array(forKey: "public_likes")!.toArray())")
+            }
+        }
+        // end::query-collection-operator-contains[]
+    }
+    
+    func dontTestCollectionOperatorIn() throws {
+        // tag::query-collection-operator-in[]
+        let values = [
+            Expression.property("first"),
+            Expression.property("last"),
+            Expression.property("username")
+        ]
+        
+        let query = QueryBuilder.select(SelectResult.all())
+            .from(DataSource.database(database))
+            .where(Expression.string("Armani").in(values))
+        // end::query-collection-operator-in[]
+        
+        print(query)
+    }
+    
+    
+    func dontTestLikeOperator() throws {
+        // tag::query-like-operator[]
+        let query = QueryBuilder
+            .select(
+                SelectResult.expression(Meta.id),
+                SelectResult.property("country"),
+                SelectResult.property("name")
+            )
+            .from(DataSource.database(database))
+            .where(Expression.property("type").equalTo(Expression.string("landmark"))
+                    .and(Function.lower(Expression.property("name"))
+                            .like(Expression.string("royal engineers museum")))
+            )
+            .limit(Expression.int(10))
+        
+        do {
+            for result in try query.execute() {
+                print("name property :: \(result.string(forKey: "name")!)")
+            }
+        }
+        // end::query-like-operator[]
+    }
+    
+    func dontTestWildCardMatch() throws {
+        // tag::query-like-operator-wildcard-match[]
+        let query = QueryBuilder
+            .select(
+                SelectResult.expression(Meta.id),
+                SelectResult.property("country"),
+                SelectResult.property("name")
+            )
+            .from(DataSource.database(database))
+            .where(Expression.property("type").equalTo(Expression.string("landmark"))
+                    .and(Function.lower(Expression.property("name"))
+                            .like(Expression.string("eng%e%")))
+            )
+            .limit(Expression.int(10))
+        // end::query-like-operator-wildcard-match[]
+        
+        do {
+            for result in try query.execute() {
+                print("name property :: \(result.string(forKey: "name")!)")
+            }
+        }
+    }
+    
+    func dontTestWildCardCharacterMatch() throws {
+        // tag::query-like-operator-wildcard-character-match[]
+        let query = QueryBuilder
+            .select(
+                SelectResult.expression(Meta.id),
+                SelectResult.property("country"),
+                SelectResult.property("name")
+            )
+            .from(DataSource.database(database))
+            .where(Expression.property("type").equalTo(Expression.string("landmark"))
+                    .and(Expression.property("name").like(Expression.string("eng____r")))
+            )
+            .limit(Expression.int(10))
+        // end::query-like-operator-wildcard-character-match[]
+        
+        do {
+            for result in try query.execute() {
+                print("name property :: \(result.string(forKey: "name")!)")
+            }
+        }
+    }
+    
+    func dontTestRegexMatch() throws {
+        // tag::query-regex-operator[]
+        let query = QueryBuilder
+            .select(
+                SelectResult.expression(Meta.id),
+                SelectResult.property("name")
+            )
+            .from(DataSource.database(database))
+            .where(Expression.property("type").equalTo(Expression.string("landmark"))
+                    .and(Expression.property("name").regex(Expression.string("\\bEng.*e\\b"))) // <.>
+            )
+            .limit(Expression.int(10))
+        // end::query-regex-operator[]
+        
+        do {
+            for result in try query.execute() {
+                print("name property :: \(result.string(forKey: "name")!)")
+            }
+        }
+    }
+    
+    func dontTestJoin() throws {
+        // tag::query-join[]
+        let query = QueryBuilder
+            .select(
+                SelectResult.expression(Expression.property("name").from("airline")),
+                SelectResult.expression(Expression.property("callsign").from("airline")),
+                SelectResult.expression(Expression.property("destinationairport").from("route")),
+                SelectResult.expression(Expression.property("stops").from("route")),
+                SelectResult.expression(Expression.property("airline").from("route"))
+            )
+            .from(
+                DataSource.database(database!).as("airline")
+            )
+            .join(
+                Join.join(DataSource.database(database!).as("route"))
+                    .on(
+                        Meta.id.from("airline")
+                            .equalTo(Expression.property("airlineid").from("route"))
+                    )
+            )
+            .where(
+                Expression.property("type").from("route").equalTo(Expression.string("route"))
+                    .and(Expression.property("type").from("airline")
+                            .equalTo(Expression.string("airline")))
+                    .and(Expression.property("sourceairport").from("route")
+                            .equalTo(Expression.string("RIX")))
+            )
+        // end::query-join[]
+        
+        do {
+            for result in try query.execute() {
+                print("name property :: \(result.string(forKey: "name")!)")
+            }
+        }
+    }
+    
+    func dontTestGroupBy() throws {
+        // tag::query-groupby[]
+        let query = QueryBuilder
+            .select(
+                SelectResult.expression(Function.count(Expression.all())),
+                SelectResult.property("country"),
+                SelectResult.property("tz"))
+            .from(DataSource.database(database))
+            .where(
+                Expression.property("type").equalTo(Expression.string("airport"))
+                    .and(Expression.property("geo.alt").greaterThanOrEqualTo(Expression.int(300)))
+            ).groupBy(
+                Expression.property("country"),
+                Expression.property("tz")
+            )
+        
+        do {
+            for result in try query.execute() {
+                print("""
+                    There are \(result.int(forKey: "$1")) airports on
+                                the \(result.string(forKey: "tz")!)timezone located
+                                in \(result.string(forKey: "country")!) and above 300 ft
+                """)
+            }
+        }
+        // end::query-groupby[]
+    }
+    
+    func dontTestOrderBy() throws {
+        // tag::query-orderby[]
+        let query = QueryBuilder
+            .select(
+                SelectResult.expression(Meta.id),
+                SelectResult.property("title"))
+            .from(DataSource.database(database))
+            .where(Expression.property("type").equalTo(Expression.string("hotel")))
+            .orderBy(Ordering.property("title").ascending())
+            .limit(Expression.int(10))
+        // end::query-orderby[]
+        
+        print("\(query)")
+    }
+    
+    func dontTestExplainAll() throws {
+        // tag::query-explain-all[]
+        let query = QueryBuilder
+            .select(SelectResult.all())
+            .from(DataSource.database(database))
+            .where(Expression.property("type").equalTo(Expression.string("university")))
+            .groupBy(Expression.property("country"))
+            .orderBy(Ordering.property("name").ascending())  // <.>
+        
+        print(try query.explain()) // <.>
+        // end::query-explain-all[]
+    }
+    
+    func dontTestExplainLike() throws {
+        // tag::query-explain-like[]
+        let query = QueryBuilder
+            .select(SelectResult.all())
+            .from(DataSource.database(database))
+            .where(Expression.property("type").like(Expression.string("%hotel%")) // <.>
+                    .and(Expression.property("name").like(Expression.string("%royal%"))));
+        
+        print(try query.explain())
+        
+        // end::query-explain-like[]
+    }
+    
+    func dontTestExplainNoOp() throws {
+        // tag::query-explain-nopfx[]
+        let query = QueryBuilder
+            .select(SelectResult.all())
+            .from(DataSource.database(database))
+            .where(Expression.property("type").like(Expression.string("hotel%")) // <.>
+                    .and(Expression.property("name").like(Expression.string("%royal%"))));
+        
+        print(try query.explain());
+        
+        // end::query-explain-nopfx[]
+    }
+    
+    func dontTestExplainFunction() throws {
+        // tag::query-explain-function[]
+        let query = QueryBuilder
+            .select(SelectResult.all())
+            .from(DataSource.database(database))
+            .where(Function.lower(Expression.property("type").equalTo(Expression.string("hotel")))) // <.>
+        
+        print(try query.explain());
+        
+        // end::query-explain-function[]
+    }
+    
+    func dontTestExplainNoFunction() throws {
+        // tag::query-explain-nofunction[]
+        let query = QueryBuilder
+            .select(SelectResult.all())
+            .from(DataSource.database(database))
+            .where(
+                Expression.property("type").equalTo(Expression.string("hotel"))); // <.>
+        
+        print(try query.explain());
+        
+        // end::query-explain-nofunction[]
+    }
+    
+    
+    func dontTestCreateFullTextIndex() throws {
+        // tag::fts-index[]
+        // Insert documents
+        let tasks = ["buy groceries", "play chess", "book travels", "buy museum tickets"]
+        for task in tasks {
+            let doc = MutableDocument()
+            doc.setString("task", forKey: "type")
+            doc.setString(task, forKey: "name")
+            try database.saveDocument(doc)
+        }
+        
+        // Create index
+        do {
+            let index = IndexBuilder.fullTextIndex(items: FullTextIndexItem.property("name")).ignoreAccents(false)
+            try database.createIndex(index, withName: "nameFTSIndex")
+        } catch let error {
+            print(error.localizedDescription)
+        }
+        // end::fts-index[]
+    }
+    
+    func dontTestFullTextSearch() throws {
+        // tag::fts-query[]
+        
+        let whereClause = FullTextFunction.match(indexName: "nameFTSIndex", query: "'buy'")
+        let query = QueryBuilder.select(SelectResult.expression(Meta.id))
+            .from(DataSource.database(database))
+            .where(whereClause)
+        
+        do {
+            for result in try query.execute() {
+                print("document id \(result.string(at: 0)!)")
+            }
+        } catch let error {
+            print(error.localizedDescription)
+        }
+        // end::fts-query[]
+    }
+    
+    // MARK: toJSON
+    
+    func dontTestToJsonArrayObject() throws {
+        // demonstrate use of JSON string
+        // tag::tojson-array[]
+        let json = "[\"1000\",\"1001\",\"1002\",\"1003\"]"
+        let myArray = try MutableArrayObject(json: json)
+        
+        for item in myArray {
+            print(item)
+        }
+        // end::tojson-array[]
+    }
+    
+    func dontTestToJsonDictionary() throws {
+        // demonstrate use of JSON string
+        // tag::tojson-dictionary[]
+        let json = "{\"id\":\"1002\",\"type\":\"hotel\",\"name\":\"Hotel Ned\"}"
+        let myDict = try MutableDictionaryObject(json: json)
+        for key in myDict {
+            if let value = myDict.value(forKey: key) {
+                print("\(key): \(value)")
+            }
+        }
+        // end::tojson-dictionary[]
+    }
+    
+    func dontTestToJsonDocument() throws {
+        // demonstrate use of JSON string
+        // tag::tojson-document[]
+        if let doc = database.document(withID: "doc-id") {
+            let json = doc.toJSON()
+            print(json)
+        }
+        // end::tojson-document[]
+    }
+    
+    func dontTestQueryResultToJSON() throws {
+        let query = QueryBuilder.select(SelectResult.all()).from(DataSource.database(database))
+        
+        // demonstrate use of JSON string
+        // tag::tojson-result[]
+        let resultSet = try query.execute()
+        for result in resultSet {
+            let json = result.toJSON()
+            print(json)
+            // end::tojson-result[]
+        }
+    }
+    
+    func dontTestBlobToJSON() throws {
+        // tag::tojson-blob[]
+        // Get a document
+        if let doc = database.document(withID: "1000") {// <.>
+            guard let blob = doc.blob(forKey: "avatar") else {
+                return
+            }
+            
+            let json = blob.toJSON() // <.>
+            print(json)
+        }
+        // end::tojson-blob[]
+        
+        // FIXME: check below function to validate whether the given property dictionary is a valid blob or not
+    }
+    
+    func dontTestIsBlob() throws {
+        let digest = ""
+        let dict = ["@type":"blob", "digest": digest]
+        
+        // tag::[dictionary-isblob]
+        if(Blob.isBlob(properties: dict)) { // <.>
+            print("Yes! I am a blob");
+        }
+        // end::[dictionary-isblob]
+    }
+    // -- !!!
+    
+    // MARK: Replication
+    
+    /* The `tag::replication[]` example is inlined in swift.adoc */
+    
+    func dontTestEnableReplicatorLogging() throws {
+        // tag::replication-logging[]
+        // Verbose / Replicator
+        Database.log.console.level = .verbose
+        Database.log.console.domains = .replicator
+        
+        // Verbose / Network
+        Database.log.console.level = .verbose
+        Database.log.console.domains = .network
+        // end::replication-logging[]
+    }
+    
+    func dontTestReplicationBasicAuthentication() throws {
+        // tag::basic-authentication[]
+        let url = URL(string: "ws://localhost:4984/mydatabase")!
+        let target = URLEndpoint(url: url)
+        var config = ReplicatorConfiguration(database: database, target: target)
+        config.authenticator = BasicAuthenticator(username: "john", password: "pass")
+        
+        self.replicator = Replicator(config: config)
+        self.replicator.start()
+        // end::basic-authentication[]
+    }
+    
+    func dontTestReplicationSessionAuthentication() throws {
+        // tag::session-authentication[]
+        let url = URL(string: "ws://localhost:4984/mydatabase")!
+        let target = URLEndpoint(url: url)
+        var config = ReplicatorConfiguration(database: database, target: target)
+        config.authenticator = SessionAuthenticator(sessionID: "904ac010862f37c8dd99015a33ab5a3565fd8447")
+        
+        self.replicator = Replicator(config: config)
+        self.replicator.start()
+        // end::session-authentication[]
+    }
+    
+    func dontTestReplicatorStatus() throws {
+        // tag::replication-status[]
+        self.replicator.addChangeListener { (change) in
+            if change.status.activity == .stopped {
+                print("Replication stopped")
+            }
+        }
+        // end::replication-status[]
+    }
+    
+    //  BEGIN PendingDocuments IB -- 11/Feb/21 --
+    func dontTestReplicationPendingDocs() throws {
+        // tag::replication-pendingdocuments[]
+        
+        let url = URL(string: "ws://localhost:4984/mydatabase")!
+        let target = URLEndpoint(url: url)
+        
+        var config = ReplicatorConfiguration(database: database, target: target)
+        config.replicatorType = .push
+        
+        // tag::replication-push-pendingdocumentids[]
+        self.replicator = Replicator(config: config)
+        let mydocids:Set = try self.replicator.pendingDocumentIds() // <.>
+        
+        // end::replication-push-pendingdocumentids[]
+        if(!mydocids.isEmpty) {
+            print("There are \(mydocids.count) documents pending")
+            let thisid = mydocids.first!
+            
+            self.replicator.addChangeListener { (change) in
+                print("Replicator activity level is \(change.status.activity)")
+                // iterate and report-on previously
+                // retrieved pending docids 'list'
+                for thisId in mydocids.sorted() {
+                    // tag::replication-push-isdocumentpending[]
+                    do {
+                        let isPending = try self.replicator.isDocumentPending(thisid)
+                        if(!isPending) { // <.>
+                            print("Doc ID \(thisId) now pushed")
+                        }
+                    } catch {
+                        print(error)
+                    }
+                    // end::replication-push-isdocumentpending[]
+                }
+            }
+            
+            self.replicator.start()
+            // end::replication-pendingdocuments[]
+        }
+    }
+    
+    //  END test PendingDocuments IB -- 11/Feb/21 --
+    
+    
+    func dontTestReplicatorDocumentEvent() throws {
+        // tag::add-document-replication-listener[]
+        let token = self.replicator.addDocumentReplicationListener { (replication) in
+            print("Replication type :: \(replication.isPush ? "Push" : "Pull")")
+            for document in replication.documents {
+                if (document.error == nil) {
+                    print("Doc ID :: \(document.id)")
+                    if (document.flags.contains(.deleted)) {
+                        print("Successfully replicated a deleted document")
+                    }
+                } else {
+                    // There was an error
+                }
+            }
+        }
+        
+        self.replicator.start()
+        // end::add-document-replication-listener[]
+        
+        // tag::remove-document-replication-listener[]
+        self.replicator.removeChangeListener(withToken: token)
+        // end::remove-document-replication-listener[]
+    }
+    
+    func dontTestReplicationCustomHeader() throws {
+        let url = URL(string: "ws://localhost:4984/mydatabase")!
+        let target = URLEndpoint(url: url)
+        
+        // tag::replication-custom-header[]
+        var config = ReplicatorConfiguration(database: database, target: target)
+        config.headers = ["CustomHeaderName": "Value"]
+        // end::replication-custom-header[]
+    }
+    
+    func dontTestReplicationChannels() throws {
+        let url = URL(string: "ws://localhost:4984/mydatabase")!
+        let target = URLEndpoint(url: url)
+        
+        // tag::replication-channels[]
+        var config = ReplicatorConfiguration(database: database, target: target)
+        config.channels = ["channel_name"]
+        // end::replication-channels[]
+    }
+    
+    func dontTestHandlingReplicationError() throws {
+        // tag::replication-error-handling[]
+        self.replicator.addChangeListener { (change) in
+            if let error = change.status.error as NSError? {
+                print("Error code :: \(error.code)")
+            }
+        }
+        // end::replication-error-handling[]
+    }
+    
+    func dontTestReplicationResetCheckpoint() throws {
+        let doResetCheckpointRequired = Bool.random()
+        
+        // tag::replication-reset-checkpoint[]
+        
+        if doResetCheckpointRequired {
+            self.replicator.start(reset: true)  // <.>
+        } else {
+            self.replicator.start()
+        }
+        
+        // end::replication-reset-checkpoint[]
+    }
+    
+    func dontTestReplicationPushFilter() throws {
+        // tag::replication-push-filter[]
+        let url = URL(string: "ws://localhost:4984/mydatabase")!
+        let target = URLEndpoint(url: url)
+        
+        var config = ReplicatorConfiguration(database: database, target: target)
+        config.pushFilter = { (document, flags) in // <1>
+            if (document.string(forKey: "type") == "draft") {
+                return false
+            }
+            return true
+        }
+        
+        self.replicator = Replicator(config: config)
+        self.replicator.start()
+        // end::replication-push-filter[]
+    }
+    
+    func dontTestReplicationPullFilter() throws {
+        // tag::replication-pull-filter[]
+        let url = URL(string: "ws://localhost:4984/mydatabase")!
+        let target = URLEndpoint(url: url)
+        
+        var config = ReplicatorConfiguration(database: database, target: target)
+        config.pullFilter = { (document, flags) in // <1>
+            if (flags.contains(.deleted)) {
+                return false
+            }
+            return true
+        }
+        
+        self.replicator = Replicator(config: config)
+        self.replicator.start()
+        // end::replication-pull-filter[]
+    }
+    
+    //  Added 2/Feb/21 - Ian Bridge
+    //  Changed for 3.0.0 - Ian Bridge 3/Mar/21
+    func testCustomRetryConfig() {
+        // tag::replication-retry-config[]
+        let target = URLEndpoint(url: URL(string: "ws://foo.couchbase.com/db")!)
+        
+        var config =  ReplicatorConfiguration(database: database, target: target)
+        config.replicatorType = .pushAndPull
+        config.continuous = true
+        // tag::replication-set-heartbeat[]
+        config.heartbeat = 150 // <.>
+        
+        // end::replication-set-heartbeat[]
+        // tag::replication-set-maxattempts[]
+        config.maxAttempts = 20 // <.>
+        
+        // end::replication-set-maxattempts[]
+        // tag::replication-set-maxattemptwaittime[]
+        config.maxAttemptWaitTime = 600 // <.>
+        self.replicator = Replicator(config: config)
+        // end::replication-set-maxattemptwaittime[]
+        
+        // end::replication-retry-config[]
+    }
+    
+#if COUCHBASE_ENTERPRISE
+    func dontTestDatabaseReplica() throws {
+        let database2 = try self.openDB(name: "db2")
+        
+        /* EE feature: code below might throw a compilation error
+         if it's compiled against CBL Swift Community. */
+        // tag::database-replica[]
+        let targetDatabase = DatabaseEndpoint(database: database2)
+        let config = ReplicatorConfiguration(database: database, target: targetDatabase)
+        config.replicatorType = .push
+        
+        self.replicator = Replicator(config: config)
+        self.replicator.start()
+        // end::database-replica[]
+        
+        try database2.delete()
+    }
+#endif
+    
+    func dontTestCertificatePinning() throws {
+        
+        // tag::certificate-pinning[]
+        let certURL = Bundle.main.url(forResource: "cert", withExtension: "cer")!
+        let data = try! Data(contentsOf: certURL)
+        let certificate = SecCertificateCreateWithData(nil, data as CFData)
+        
+        let url = URL(string: "wss://localhost:4985/db")!
+        let target = URLEndpoint(url: url)
+        
+        var config = ReplicatorConfiguration(database: database, target: target)
+        config.pinnedServerCertificate = certificate
+        // end::certificate-pinning[]
+        
+        print("\(config)")
+    }
+    
+    func dontTestGettingStarted() throws {
+        // tag::getting-started[]
+        // Get the database (and create it if it doesn’t exist).
+        let database: Database
+        do {
+            database = try Database(name: "mydb")
+        } catch {
+            fatalError("Error opening database")
+        }
+        
+        // Create a new document (i.e. a record) in the database.
+        let mutableDoc = MutableDocument()
+            .setFloat(2.0, forKey: "version")
+            .setString("SDK", forKey: "type")
+        
+        // Save it to the database.
+        do {
+            try database.saveDocument(mutableDoc)
+        } catch {
+            fatalError("Error saving document")
+        }
+        
+        // Update a document.
+        if let mutableDoc = database.document(withID: mutableDoc.id)?.toMutable() {
+            mutableDoc.setString("Swift", forKey: "language")
+            do {
+                try database.saveDocument(mutableDoc)
+                
+                let document = database.document(withID: mutableDoc.id)!
+                // Log the document ID (generated by the database)
+                // and properties
+                print("Document ID :: \(document.id)")
+                print("Learning \(document.string(forKey: "language")!)")
+            } catch {
+                fatalError("Error updating document")
+            }
+        }
+        
+        // Create a query to fetch documents of type SDK.
+        let query = QueryBuilder
+            .select(SelectResult.all())
+            .from(DataSource.database(database))
+            .where(Expression.property("type").equalTo(Expression.string("SDK")))
+        
+        // Run the query.
+        do {
+            let result = try query.execute()
+            print("Number of rows :: \(result.allResults().count)")
+        } catch {
+            fatalError("Error running the query")
+        }
+        
+        // Create replicators to push and pull changes to and from the cloud.
+        let targetEndpoint = URLEndpoint(url: URL(string: "ws://localhost:4984/getting-started-db")!)
+        var replConfig = ReplicatorConfiguration(database: database, target: targetEndpoint)
+        replConfig.replicatorType = .pushAndPull
+        
+        // Add authentication.
+        replConfig.authenticator = BasicAuthenticator(username: "john", password: "pass")
+        
+        // Create replicator (make sure to add an instance or static variable named replicator)
+        self.replicator = Replicator(config: replConfig)
+        
+        // Listen to replicator change events.
+        self.replicator.addChangeListener { (change) in
+            if let error = change.status.error as NSError? {
+                print("Error code :: \(error.code)")
+            }
+        }
+        
+        // Start replication.
+        self.replicator.start()
+        // end::getting-started[]
+    }
+    
+    func dontTestPredictiveModel() throws {
+        let database: Database
+        do {
+            database = try Database(name: "mydb")
+        } catch {
+            fatalError("Error opening database")
+        }
+        
+        // tag::register-model[]
+        let model = ImageClassifierModel()
+        Database.prediction.registerModel(model, withName: "ImageClassifier")
+        // end::register-model[]
+        
+        // tag::predictive-query-value-index[]
+        let input = Expression.dictionary(["photo": Expression.property("photo")])
+        let prediction = Function.prediction(model: "ImageClassifier", input: input)
+        
+        let index = IndexBuilder.valueIndex(items: ValueIndexItem.expression(prediction.property("label")))
+        try database.createIndex(index, withName: "value-index-image-classifier")
+        // end::predictive-query-value-index[]
+        
+        // tag::unregister-model[]
+        Database.prediction.unregisterModel(withName: "ImageClassifier")
+        // end::unregister-model[]
+    }
+    
+    func dontTestPredictiveIndex() throws {
+        let database: Database
+        do {
+            database = try Database(name: "mydb")
+        } catch {
+            fatalError("Error opening database")
+        }
+        
+        // tag::predictive-query-predictive-index[]
+        let input = Expression.dictionary(["photo": Expression.property("photo")])
+        
+        let index = IndexBuilder.predictiveIndex(model: "ImageClassifier", input: input)
+        try database.createIndex(index, withName: "predictive-index-image-classifier")
+        // end::predictive-query-predictive-index[]
+    }
+    
+    func dontTestPredictiveQuery() throws {
+        let database: Database
+        do {
+            database = try Database(name: "mydb")
+        } catch {
+            fatalError("Error opening database")
+        }
+        
+        // tag::predictive-query[]
+        let input = Expression.dictionary(["photo": Expression.property("photo")])
+        let prediction = Function.prediction(model: "ImageClassifier", input: input) // <1>
+        
+        let query = QueryBuilder
+            .select(SelectResult.all())
+            .from(DataSource.database(database))
+            .where(
+                prediction.property("label").equalTo(Expression.string("car"))
+                    .and(
+                        prediction.property("probablity")
+                            .greaterThanOrEqualTo(Expression.double(0.8))
+                    )
+            )
+        
+        // Run the query.
+        do {
+            let result = try query.execute()
+            print("Number of rows :: \(result.allResults().count)")
+        } catch {
+            fatalError("Error running the query")
+        }
+        // end::predictive-query[]
+    }
+    
+    func dontTestCoreMLPredictiveModel() throws {
+        // tag::coreml-predictive-model[]
+        // Load MLModel from `ImageClassifier.mlmodel`
+        let modelURL = Bundle.main.url(forResource: "ImageClassifier", withExtension: "mlmodel")!
+        let compiledModelURL = try MLModel.compileModel(at: modelURL)
+        let model = try MLModel(contentsOf: compiledModelURL)
+        let predictiveModel = CoreMLPredictiveModel(mlModel: model)
+        
+        // Register model
+        Database.prediction.registerModel(predictiveModel, withName: "ImageClassifier")
+        // end::coreml-predictive-model[]
+    }
+    
+    func dontTestReplicatorConflictResolver() throws {
+        // tag::replication-conflict-resolver[]
+        let url = URL(string: "wss://localhost:4984/mydatabase")!
+        let target = URLEndpoint(url: url)
+        
+        var config = ReplicatorConfiguration(database: database, target: target)
+        config.conflictResolver = LocalWinConflictResolver()
+        
+        self.replicator = Replicator(config: config)
+        self.replicator.start()
+        // end::replication-conflict-resolver[]
+    }
+    
+    func dontTestSaveWithConflictHandler() throws {
+        // tag::update-document-with-conflict-handler[]
+        guard let document = database.document(withID: "xyz") else { return }
+        let mutableDocument = document.toMutable()
+        mutableDocument.setString("apples", forKey: "name")
+        try database.saveDocument(mutableDocument, conflictHandler: { (new, current) -> Bool in
+            let currentDict = current!.toDictionary()
+            let newDict = new.toDictionary()
+            let result = newDict.merging(currentDict, uniquingKeysWith: { (first, _) in first })
+            new.setData(result)
+            return true
+        })
+        // end::update-document-with-conflict-handler[]
+        
+    }
+    
+    // helper
+    func isValidCredentials(_ u: String, password: String) -> Bool { return true }
+    
+    func dontTestInitListener() throws {
+        // tag::init-urllistener[]
+        var config = URLEndpointListenerConfiguration(database: otherDB)
+        config.tlsIdentity = nil; // Use with anonymous self signed cert
+        config.authenticator = ListenerPasswordAuthenticator(authenticator: { (username, password) -> Bool in
+            return self.isValidCredentials(username, password: password)
+        })
+        
+        // end::init-urllistener[]
+    }
+    
+    func dontTestListenerStart() throws {
+        // tag::start-urllistener[]
+        try listener.start()
+        
+        // end::start-urllistener[]
+    }
+    
+    func dontTestListenerStop() throws {
+        // tag::stop-urllistener[]
+        listener.stop()
+        
+        // end::stop-urllistener[]
+    }
+    
+    func dontTestCreateSelfSignedCert() throws {
+        // <site-rooot>/objc/advance/objc-p2psync-websocket-using-passive.html
+        // Example-6
+        // tag::create-self-signed-cert[]
+        // tag::listener-config-tls-id-SelfSigned[]
+        let attrs = [certAttrCommonName: "Couchbase Inc"]
+        let identity =
+        try TLSIdentity.createIdentity(forServer: true,
+                                       attributes: attrs,
+                                       expiration: Date().addingTimeInterval(86400),
+                                       label: "Server-Cert-Label")
+        // end::listener-config-tls-id-SelfSigned[]
+        // end::create-self-signed-cert[]
+        print("\(identity.expiration)") // to avoid warning
+    }
+    
+    // MARK: -- URLEndpointListener
+    func dontTestTLSIdentityCreate() throws {
+        // tag::p2psync-act-tlsid-create[]
+        try TLSIdentity.deleteIdentity(withLabel: "couchbaselite-server-cert-label");
+        let attrs = [certAttrCommonName: "CBL-Server"]
+        let identity = try TLSIdentity.createIdentity(forServer: true,
+                                                      attributes: attrs,
+                                                      expiration: nil,
+                                                      label: "couchbaselite-server-cert-label")
+        // end::p2psync-act-tlsid-create[]
+        print(identity)
+    }
+    
+    func dontTestDeleteIdentity() throws {
+        // tag::p2psync-act-tlsid-delete[]
+        try TLSIdentity.deleteIdentity(withLabel: "couchbaselite-client-cert-label")
+        // end::p2psync-act-tlsid-delete[]
+    }
+    
+    func dontTestImportTLSIdentity() throws {
+        // tag::p2psync-act-tlsid-import[]
+        
+        let path = Bundle.main.path(forResource: "identity/client", ofType: "p12")
+        let clientCertData = try NSData(contentsOfFile: path!, options: []) as Data
+        let identity = try TLSIdentity.importIdentity(withData: clientCertData,
+                                                      password: "123",
+                                                      label: "couchbaselite-client-cert-label")
+        // end::p2psync-act-tlsid-import[]
+        print(identity)
+    }
+    
+    // MARK: -- QUERY RESULT SET HANDLING EXAMPLES
+    
+    func donTestQuerySyntaxAll() throws {
+        // tag::query-syntax-all[]
+        let database = try Database(name: "hotel")
+        let query = QueryBuilder.select(SelectResult.all()).from(DataSource.database(database))
+        
+        // end::query-syntax-all[]
+        
+        print(query)
+    }
+    
+    func dontTestQueryAccessAll() throws {
+        let query = QueryBuilder.select(SelectResult.all()).from(DataSource.database(database))
+        var hotels = [String: Any]()
+        
+        // tag::query-access-all[]
+        let results = try query.execute()
+        for row in results {
+            let docsProps = row.dictionary(at: 0)?.toDictionary() // <.>
+            
+            let docid = docsProps!["id"] as! String
+            let name = docsProps!["name"] as! String
+            let type = docsProps!["type"] as! String
+            let city = docsProps!["city"] as! String
+            
+            print("\(docid): \(name), \(type), \(city)")
+            let hotel = row.dictionary(at: 0)?.toDictionary()  //<.>
+            let hotelId = hotel!["id"] as! String
+            
+            hotels[hotelId] = hotel
+        }
+        
+        // end::query-access-all[]
+    }
+    
+    
+    
+    func dontTestQueryAccessJSON() throws {
+        let query = QueryBuilder.select(SelectResult.all()).from(DataSource.database(database))
+        var hotels = [String:Hotel]()
+        // tag::query-access-json[]
+        
+        // In this example the Hotel class is defined using Codable
+        //
+        // class Hotel : Codable {
+        //   var id : String = "undefined"
+        //   var type : String = "hotel"
+        //   var name : String = "undefined"
+        //   var city : String = "undefined"
+        //   var country : String = "undefined"
+        //   var description : String? = ""
+        //   var text : String? = ""
+        //   ... other class content
+        // }
+        
+        let results = try query.execute()
+        for row in  results {
+            
+            // get the result into a JSON String
+            let jsonString = row.toJSON() // <.>
+            
+            let thisJsonObj:Dictionary =
+            try (JSONSerialization.jsonObject(
+                with: jsonString.data(using: .utf8)!,
+                options: .allowFragments)
+                 as? [String: Any])! // <.>
+            
+            // Use Json Object to populate Native object
+            // Use Codable class to unpack JSON data to native object
+            var this_hotel: Hotel = try JSONDecoder().decode(Hotel.self, from: jsonString.data(using: .utf8)!) // <.>
+            
+            // ALTERNATIVELY unpack in steps
+            this_hotel.id = thisJsonObj["id"] as! String
+            this_hotel.name = thisJsonObj["name"] as! String
+            this_hotel.type = thisJsonObj["type"] as! String
+            this_hotel.city = thisJsonObj["city"] as! String
+            hotels[this_hotel.id] = this_hotel
+            
+        } // end for
+        
+        // end::query-access-json[]
+    }
+    
+    func dontTestQuerySyntaxProps() throws {
+        // tag::query-syntax-props[]
+        let database = try! Database(name: "hotel")
+        
+        let query = QueryBuilder
+            .select(SelectResult.expression(Meta.id).as("metaId"),
+                    SelectResult.expression(Expression.property("id")),
+                    SelectResult.expression(Expression.property("name")),
+                    SelectResult.expression(Expression.property("city")),
+                    SelectResult.expression(Expression.property("type")))
+            .from(DataSource.database(database))
+        
+        // end::query-syntax-props[]
+        print(query)
+    }
+    
+    func dontTestQueryAccessProps () throws {
+        let query = QueryBuilder.select(SelectResult.all()).from(DataSource.database(database))
+        var hotels = [String: Hotel]()
+        
+        // tag::query-access-props[]
+        for result in try! query.execute() {
+            
+            
+            let doc = result.toDictionary()  // <.>
+            
+            // Store dictionary data in hotel object and save in arry
+            guard let id = doc["id"] as? String else {
+                continue
+            }
+            var hotel = Hotel(id: id)
+            hotel.name = doc["name"] as? String
+            hotel.city = doc["city"] as? String
+            hotel.type = doc["type"] as? String
+            hotels[id] = hotel
+            
+            // Use result content directly
+            let docid = result.string(forKey: "metaId") // Selected (Meta.id).as("metaId")
+            let hotelId = result.string(forKey: "id")
+            let name = result.string(forKey: "name")
+            let city = result.string(forKey: "city")
+            let type = result.string(forKey: "type")
+            
+            // ... process document properties as required
+            print("Result properties are: ", docid, hotelId,name, city, type)
+        } // end for
+        
+        // end::query-access-props[]
+    }// end func
+    
+    func dontTestQueryCount() throws {
+        // tag::query-syntax-count-only[]
+        let database = try Database(name: "hotel")
+        let query = QueryBuilder
+            .select(SelectResult.expression(Function.count(Expression.all())).as("mycount"))
+            .from (DataSource.database(database)).groupBy(Expression.property("type"))
+        
+        // end::query-syntax-count-only[]
+        
+        // tag::query-access-count-only[]
+        for result in try query.execute() {
+            let dict = result.toDictionary() as? [String: Int]
+            let thiscount = dict!["mycount"]! // <.>
+            print("There are ", thiscount, " rows")
+            
+            // Alternatively
+            print ( result["mycount"] )
+            
+        }
+        // end::query-access-count-only[]
+    }
+    
+    func dontTestQueryId () throws {
+        // tag::query-syntax-id[]
+        let database = try Database(name: "hotel")
+        let query = QueryBuilder.select(SelectResult.expression(Meta.id).as("metaId"))
+            .from(DataSource.database(database))
+        
+        // end::query-syntax-id[]
+        
+        
+        // tag::query-access-id[]
+        let results = try query.execute()
+        for result in results {
+            
+            print(result.toDictionary())
+            print("Document Id is -- ", result["metaId"].string!)
+            
+            let docId = result["metaId"].string! // <.>
+            
+            // Now you can get the document using the ID
+            let doc = database.document(withID: docId)!.toDictionary()
+            
+            let hotelId = doc["id"] as! String
+            
+            let name = doc["name"] as! String
+            
+            let city = doc["city"] as! String
+            
+            let type = doc["type"] as! String
+            
+            // ... process document properties as required
+            print("Result properties are: ", hotelId,name, city, type)
+            
+            
+        }
+        // end::query-access-id[]
+    }
+    
+    func query_pagination () throws {
+        //tag::query-syntax-pagination[]
+        let offset = 0;
+        let limit = 20;
+        //
+        let query = QueryBuilder
+            .select(SelectResult.all())
+            .from(DataSource.database(database))
+            .limit(Expression.int(limit), offset: Expression.int(offset))
+        
+        // end::query-syntax-pagination[]
+        print(query)
+    }
+    
+    func dontTestQueryN1QL() throws {
+        
+        // tag::query-syntax-n1ql[]
+        let database = try Database(name: "hotel")
+        
+        let query =  database.createQuery(query: "SELECT META().id AS thisId FROM _ WHERE type = 'hotel'") // <.>
+        
+        let results: ResultSet = try query.execute()
+        
+        // end::query-syntax-n1ql[]
+        
+        print(results.allResults().count)
+    }
+    
+    func dontTestQueryN1QLparams() throws {
+        
+        // tag::query-syntax-n1ql-params[]
+        let database = try! Database(name: "hotel")
+        
+        let query =
+        database.createQuery(query: "SELECT META().id AS thisId FROM _ WHERE type = $type") // <.>
+        
+        query.parameters =
+        Parameters().setString("hotel", forName: "type") // <.>
+        
+        let results: ResultSet = try query.execute()
+        
+        // end::query-syntax-n1ql-params[]
+        
+        print(results.allResults().count)
+    }
+    
+    func dontTestProcessResults(results: ResultSet) throws {
+        // tag::query-access-n1ql[]
+        // tag::query-process-results[]
+        
+        for row in results {
+            print(row["thisId"].string!)
+            
+            let docsId = row["thisId"].string!
+            
+            // Now you can get the document using the ID
+            let doc = database.document(withID: docsId)!.toDictionary()
+            
+            let hotelId = doc["id"] as! String
+            
+            let name = doc["name"] as! String
+            
+            let city = doc["city"] as! String
+            
+            let type = doc["type"] as! String
+            
+            // ... process document properties as required
+            print("Result properties are: ", hotelId,name, city, type)
+            
+        }
+        // end::query-access-n1ql[]
+        // end::query-process-results[]
+        
+    }
+    
+    // MARK: -- Listener
+    
+    func dontTestListenerSimple() throws {
+        let database = try Database(name: "database")
+        // tag::listener-simple[]
+        var config = URLEndpointListenerConfiguration(database: database) // <.>
+        config.authenticator = ListenerPasswordAuthenticator { username, password in
+            return "valid.user" == username && "valid.password.string" == String(password)
+        } // <.>
+        
+        let listener = URLEndpointListener(config: config) // <.>
+        
+        try listener.start()  // <.>
+        
+        // end::listener-simple[]
+    }
+    
+    func dontTestListenerInitialize() throws {
+        let otherDB = try Database(name: "otherDB")
+        // tag::listener-initialize[]
+        
+        // tag::listener-config-db[]
+        var config = URLEndpointListenerConfiguration(database: otherDB) // <.>
+        
+        // end::listener-config-db[]
+        // tag::listener-config-port[]
+        /* optionally */ let wsPort: UInt16 = 55991
+        /* optionally */ let wssPort: UInt16 = 55990
+        config.port =  wssPort // <.>
+        
+        // end::listener-config-port[]
+        // tag::listener-config-netw-iface[]
+        config.networkInterface = "10.1.1.10"  // <.>
+        
+        // end::listener-config-netw-iface[]
+        // tag::listener-config-delta-sync[]
+        config.enableDeltaSync = true // <.>
+        
+        // end::listener-config-delta-sync[]
+        // tag::listener-config-tls-enable[]
+        config.disableTLS  = false // <.>
+        
+        // end::listener-config-tls-enable[]
+        // tag::listener-config-tls-id-anon[]
+        // Set the credentials the server presents the client
+        // Use an anonymous self-signed cert
+        config.tlsIdentity = nil // <.>
+        
+        // end::listener-config-tls-id-anon[]
+        // tag::listener-config-client-auth-pwd[]
+        // Configure how the client is to be authenticated
+        // Here, use Basic Authentication
+        config.authenticator = ListenerPasswordAuthenticator(authenticator: { uname, pword -> Bool in
+            return self.isValidCredentials(uname, password: pword)
+        }) // <.>
+        
+        // end::listener-config-client-auth-pwd[]
+        
+        // tag::listener-start[]
+        // Initialize the listener
+        self.listener = URLEndpointListener(config: config) // <.>
+        if self.listener == nil {
+            fatalError("ListenerError Not Initialized")
+            // ... take appropriate actions
+        }
+        
+        // Start the listener
+        try self.listener.start() // <.>
+        
+        // end::listener-start[]
+        // end::listener-initialize[]
+        
+        print(wsPort)
+    }
+    
+    func dontTestReplicatorSimple() throws {
+        // tag::replicator-simple[]
+        
+        let tgtUrl = URL(string: "wss://10.1.1.12:8092/otherDB")!
+        let targetEndpoint = URLEndpoint(url: tgtUrl) //  <.>
+        
+        var thisConfig = ReplicatorConfiguration(database: database, target: targetEndpoint) // <.>
+        
+        thisConfig.acceptOnlySelfSignedServerCertificate = true // <.>
+        
+        let thisAuthenticator = BasicAuthenticator(username: "valid.user", password: "valid.password.string")
+        thisConfig.authenticator = thisAuthenticator // <.>
+        
+        self.replicator = Replicator(config: thisConfig) // <.>
+        
+        self.replicator.start(); // <.>
+        
+        // end::replicator-simple[]
+    }
+    
+    // FIXME: Could you please check, whether this is necessary? it was a unit test helper function!
+    // tag::xctListener-start-func[]
+    func listen(tls: Bool, auth: ListenerAuthenticator?) throws -> URLEndpointListener {
+        let otherDB = try? Database(name: "thisDB")
+        let wssPort: UInt16 = 4985
+        let wsPort: UInt16 = 4984
+        // Stop if already running :
+        if let listener = self.listener {
+            listener.stop()
+        }
+        
+        // Listener:
+        // tag::xctListener-start[]
+        // tag::xctListener-config[]
+        //  ... fragment preceded by other user code, including
+        //  ... Couchbase Lite Database initialization that returns `thisDB`
+        
+        guard let otherDB = otherDB else {
+            fatalError("DatabaseNotInitialized")
+            // ... take appropriate actions
+        }
+        
+        var config = URLEndpointListenerConfiguration.init(database: otherDB)
+        config.port = tls ? wssPort : wsPort
+        config.disableTLS = !tls
+        config.authenticator = auth
+        listener = URLEndpointListener.init(config: config)
+        //  ... fragment followed by other user code
+        // end::xctListener-config[]
+        
+        // Start:
+        try listener.start()
+        // end::xctListener-start[]
+        
+        return listener
+    }
+    // end::xctListener-start-func[]
+    
+    
+    func dontTestStopListener(listener: URLEndpointListener) throws {
+        // tag::xctListener-stop-func[]
+        listener.stop()
+        // end::xctListener-stop-func[]
+        
+        // FIXME: Removed an internal only function 'deleteFromKeyChain'; Not public API to expose!
+    }
+    
+    // FIXME: this is only for tests: Internal only: Don't expose this!!
+    //    func dontTestDeleteAnonymousIds() throws {
+    // tag::xctListener-delete-anon-ids[]
+    //        FIXME: try URLEndpointListener.deleteAnonymousIdentities()
+    // end::xctListener-delete-anon-ids[]
+    //    }
+    
+    func dontTestTLSIdentityAnonym() throws {
+        // tag::xctListener-auth-tls-tlsidentity-anon[]
+        // Anonymous Identity:
+        let config = URLEndpointListenerConfiguration.init(database: database)
+        self.listener = URLEndpointListener.init(config: config)
+        try self.listener.start()
+        
+        
+        print(self.listener.tlsIdentity!.certs) // Anonymous `self.listener.tlsIdentity`
+        // end::xctListener-auth-tls-tlsidentity-anon[]
+    }
+    
+    func testTLSIdentityCA() throws {
+        // tag::xctListener-auth-tls-tlsidentity-ca[]
+        // User Identity:
+        let attrs = [certAttrCommonName: "CBL-Server"]
+        let identity = try TLSIdentity.createIdentity(forServer: true,
+                                                      attributes: attrs,
+                                                      expiration: nil,
+                                                      label: "couchbaselite-server-cert-label")
+        var config = URLEndpointListenerConfiguration.init(database: otherDB)
+        config.tlsIdentity = identity
+        self.listener = URLEndpointListener.init(config: config)
+        
+        try self.listener.start()
+        print(self.listener.tlsIdentity!.certs) // 'couchbaselite-server-cert-label' tlsIdentity
+        // end::xctListener-auth-tls-tlsidentity-ca[]
+    }
+    
+    func testPasswordAuthenticator() throws {
+        // tag::xctListener-auth-basic-pwd-full[]
+        // Listener:
+        // tag::xctListener-auth-basic-pwd[]
+        var config = URLEndpointListenerConfiguration(database: otherDB)
+        config.authenticator = ListenerPasswordAuthenticator.init { username, password -> Bool in
+            return username == "daniel" && password == "123"
+        }
+        
+        self.listener = URLEndpointListener(config: config)
+        try self.listener.start()
+        // end::xctListener-auth-basic-pwd[]
+    }
+    // end::xctListener-auth-basic-pwd-full[]
+    
+    func testClientCertAuthenticatorWithRootCerts() throws {
+        // tag::xctListener-auth-tls-CCA-Root-full[]
+        // tag::xctListener-auth-tls-CCA-Root[]
+        // Root Cert:
+        let path = Bundle.main.path(forResource: "identity/client-ca", ofType: "der")
+        let rootCertData = try NSData(contentsOfFile: path!, options: []) as Data
+        let rootCert = SecCertificateCreateWithData(kCFAllocatorDefault, rootCertData as CFData)!
+        
+        // Listener:
+        var config = URLEndpointListenerConfiguration(database: otherDB)
+        config.authenticator = ListenerCertificateAuthenticator.init(rootCerts: [rootCert])
+        self.listener = URLEndpointListener(config: config)
+        
+        try self.listener.start()
+        // end::xctListener-auth-tls-CCA-Root[]
+        
+        
+        // Create client identity:
+        let clientCertPath = Bundle.main.path(forResource: "identity/client", ofType: "p12")
+        let clientCertData = try NSData(contentsOfFile: clientCertPath!, options: []) as Data
+        let identity = try TLSIdentity.importIdentity(withData: clientCertData,
+                                                      password: "123",
+                                                      label: "couchbaselite-client-cert-label")
+        
+        // Replicator:
+        let target = URLEndpoint(url: URL(string: "wss://localhost:4985/otherDB")!)
+        var replicatorConfig = ReplicatorConfiguration(database: database, target: target)
+        replicatorConfig.replicatorType = .pushAndPull
+        replicatorConfig.continuous = false
+        replicatorConfig.authenticator = ClientCertificateAuthenticator.init(identity: identity)
+        replicatorConfig.pinnedServerCertificate = self.listener.tlsIdentity!.certs[0]
+        self.replicator = Replicator(config: replicatorConfig)
+        
+        self.replicator.start()
+        // end::xctListener-auth-tls-CCA-Root-full[]
+    }
+    
+    func testServerCertVerificationModeSelfSignedCert() throws {
+        // tag::xctListener-auth-tls-self-signed-full[]
+        // tag::xctListener-auth-tls-self-signed[]
+        // Listener:
+        let config = URLEndpointListenerConfiguration(database: otherDB)
+        self.listener = URLEndpointListener(config: config)
+        try self.listener.start()
+        
+        print(self.listener.tlsIdentity!.certs) // Anonymous `self.listener.tlsIdentity`
+        // end::xctListener-auth-tls-self-signed[]
+        
+        // Connect replicator with accept only self signed config
+        let target = URLEndpoint(url: URL(string: "wss://localhost:4985/otherDB")!)
+        var replicatorConfig = ReplicatorConfiguration(database: database, target: target)
+        replicatorConfig.replicatorType = .pushAndPull
+        replicatorConfig.continuous = false
+        replicatorConfig.acceptOnlySelfSignedServerCertificate = true
+        self.replicator = Replicator(config: replicatorConfig)
+        self.replicator.start()
+        
+        // end::xctListener-auth-tls-self-signed-full[]
+    }
+    
+    // tag::xctListener-auth-tls-ca-cert-full[]
+    func testServerCertVerificationModeCACert() throws {
+        // Listener:
+        // tag::xctListener-auth-tls-ca-cert[]
+        let rootCertPath = Bundle.main.path(forResource: "identity/certs", ofType: "p12")
+        let rootCertData = try NSData(contentsOfFile: rootCertPath!, options: []) as Data
+        var config = URLEndpointListenerConfiguration(database: otherDB)
+        let identity = try TLSIdentity.importIdentity(withData: rootCertData,
+                                                      password: "123",
+                                                      label: "couchbaselite-server-cert-label")
+        config.tlsIdentity = identity
+        self.listener = URLEndpointListener(config: config)
+        try self.listener.start()
+        
+        print(self.listener.tlsIdentity!.certs) // `self.listener.tlsIdentity` TLSIdentity with CA
+        
+        // Replicator
+        let target = URLEndpoint(url: URL(string: "wss://localhost:4985/otherDB")!)
+        var replicatorConfig = ReplicatorConfiguration(database: database, target: target)
+        replicatorConfig.replicatorType = .pushAndPull
+        replicatorConfig.continuous = false
+        replicatorConfig.pinnedServerCertificate = self.listener.tlsIdentity!.certs[0]
+        self.replicator = Replicator(config: replicatorConfig)
+        self.replicator.start()
+        // end::xctListener-auth-tls-ca-cert[]
+    }
+    
+    // tag::xctListener-status-check-full[]
+    // FIXME: Can we link to the unit test??
+    // https://github.com/couchbase/couchbase-lite-ios/blob/release/lithium/Swift/Tests/URLEndpointListenerTest.swift#L626-L676
+    // ??
+    // end::xctListener-status-check-full[]
+    
+    func dontTestListenerBasicAuthPassword() throws {
+        var config = URLEndpointListenerConfiguration(database: otherDB)
+        // tag::xctListener-auth-password-basic[]
+        config.authenticator = ListenerPasswordAuthenticator { self.isValidCredentials($0, password: $1) }
+        // end::xctListener-auth-password-basic[]
+    }
+    
+    func dontTestListenerAuthCertRoots() throws {
+        // tag::xctListener-auth-cert-roots[]
+        let path = Bundle.main.path(forResource: "identity/client-ca", ofType: "der")
+        let rootCertData = try NSData(contentsOfFile: path!, options: []) as Data
+        let rootCert = SecCertificateCreateWithData(kCFAllocatorDefault, rootCertData as CFData)!
+        
+        var config = URLEndpointListenerConfiguration(database: otherDB)
+        config.authenticator = ListenerCertificateAuthenticator.init(rootCerts: [rootCert])
+        listener = URLEndpointListener(config: config)
+        // end::xctListener-auth-cert-roots[]
+        
+        // tag::xctListener-auth-cert-auth[]
+        let listenerAuth = ListenerCertificateAuthenticator { self.isValidCertificates($0) }
+        // end::xctListener-auth-cert-auth[]
+        print(listenerAuth)
+    }
+    
+    func dontTestBasicAuth() throws {
+        var config = URLEndpointListenerConfiguration(database: otherDB)
+        
+        // tag::xctListener-config-basic-auth[]
+        config = URLEndpointListenerConfiguration(database: otherDB)
+        
+        config.authenticator = ListenerPasswordAuthenticator { self.isValidCredentials($0, password: $1) }
+        
+        self.listener = URLEndpointListener(config: config)
+        // end::xctListener-config-basic-auth[]
+    }
+    
+    // MARK: Append
+    
+    func dontTestGetURLList() throws {
+        // tag::listener-get-url-list[]
+        let config = URLEndpointListenerConfiguration(database: otherDB)
+        let listener = URLEndpointListener(config: config)
+        try listener.start()
+        
+        if let urls = listener.urls {
+            print("URLs are: \(urls)")
+        }
+        
+        // end::listener-get-url-list[]
+    }
+    
+    
+    func dontTestListenerConfigDisableTLSUpdate() throws {
+        var config = URLEndpointListenerConfiguration(database: otherDB)
+        // tag::listener-config-tls-full-enable[]
+        config.disableTLS  = false // <.>
+        
+        // end::listener-config-tls-full-enable[]
+        // tag::listener-config-tls-disable[]
+        config.disableTLS  = true // <.>
+        
+        // end::listener-config-tls-disable[]
+    }
+    
+    func dontTestListenerConfigTLSIdentity() throws {
+        var config = URLEndpointListenerConfiguration(database: otherDB)
+        
+        // tag::listener-config-tls-id-full[]
+        // tag::listener-config-tls-id-caCert[]
+        guard let path = Bundle.main.path(forResource: "cert", ofType: "p12") else {
+            /* process error */ return
+        }
+        
+        guard let certData = try? NSData(contentsOfFile: path) as Data else {
+            /* process error */ return
+        } // <.>
+        
+        let tlsIdentity = try TLSIdentity.importIdentity(withData: certData,
+                                                         password: "123",
+                                                         label: "Server-Cert-Label") // <.>
+        
+        // end::listener-config-tls-id-caCert[]
+        // tag::listener-config-tls-id-SelfSigned[]
+        let attrs = [certAttrCommonName: "Couchbase Inc"] // <.>
+        
+        let identity = try TLSIdentity.createIdentity(forServer: true, /* isServer */
+                                                      attributes: attrs,
+                                                      expiration: Date().addingTimeInterval(86400),
+                                                      label: "Server-Cert-Label") // <.>
+        
+        // end::listener-config-tls-id-SelfSigned[]
+        // tag::listener-config-tls-id-full-set[]
+        // Set the credentials the server presents the client
+        config.tlsIdentity = tlsIdentity    // <.>
+        
+        // end::listener-config-tls-id-full-set[]
+        // end::listener-config-tls-id-full[]
+        
+        print("To avoid waring: \(identity)")
+    }
+    
+    func dontTestListenerConfigClientRootCA() throws {
+        var config = URLEndpointListenerConfiguration(database: otherDB)
+        let cert = self.listener.tlsIdentity!.certs[0]
+        
+        // tag::listener-config-client-root-ca[]
+        // tag::listener-config-client-auth-root[]
+        // Authenticate using Cert Authority
+        
+        // cert is a pre-populated object of type:SecCertificate representing a certificate
+        let rootCertData = SecCertificateCopyData(cert) as Data // <.>
+        let rootCert = SecCertificateCreateWithData(kCFAllocatorDefault, rootCertData as CFData)! //
+        
+        config.authenticator = ListenerCertificateAuthenticator.init (rootCerts: [rootCert]) // <.> <.>
+        
+        // end::listener-config-client-auth-root[]
+        // end::listener-config-client-root-ca[]
+    }
+    
+    func isValidCertificates(_ certs: [SecCertificate]) -> Bool { return true }
+    
+    func dontTestClientAuthLambda() throws {
+        var config = URLEndpointListenerConfiguration(database: otherDB)
+        
+        // tag::listener-config-client-auth-lambda[]
+        // tag::listener-config-client-auth-self-signed[]
+        // Authenticate self-signed cert using application logic
+        
+        config.authenticator = ListenerCertificateAuthenticator { certs -> Bool in // <.>
+            // Validate the cert
+            return self.isValidCertificates(certs)
+        } // <.>
+        
+        // end::listener-config-client-auth-self-signed[]
+        // end::listener-config-client-auth-lambda[]
+    }
+    
+    func dontTestListenerConfigUpdate() throws {
+        var config = URLEndpointListenerConfiguration(database: otherDB)
+        let cert = self.listener.tlsIdentity!.certs[0]
+        
+        // tag::old-listener-config-tls-id-nil[]
+        config.tlsIdentity = nil
+        
+        // end::old-listener-config-tls-id-nil[]
+        // tag::old-listener-config-delta-sync[]
+        config.enableDeltaSync = true
+        
+        // end::old-listener-config-delta-sync[]
+        // tag::listener-status-check[]
+        let totalConnections = self.listener.status.connectionCount
+        let activeConnections = self.listener.status.activeConnectionCount
+        
+        // end::listener-status-check[]
+        // tag::listener-stop[]
+        self.listener.stop()
+        
+        // end::listener-stop[]
+        // tag::old-listener-config-client-auth-root[]
+        // cert is a pre-populated object of type:SecCertificate representing a certificate
+        let rootCertData = SecCertificateCopyData(cert) as Data
+        let rootCert = SecCertificateCreateWithData(kCFAllocatorDefault, rootCertData as CFData)!
+        
+        config.authenticator = ListenerCertificateAuthenticator.init (rootCerts: [rootCert])
+        
+        // end::old-listener-config-client-auth-root[]
+        // tag::old-listener-config-client-auth-self-signed[]
+        config.authenticator = ListenerCertificateAuthenticator { self.isValidCertificates($0) }
+        // end::old-listener-config-client-auth-self-signed[]
+        
+        print("to avoid warnings: \(activeConnections)/\(totalConnections) ")
+    }
+    
+    func dontTestURLEndpointListenerConstructor() throws {
+        let enableTLS = Bool.random()
+        let wssPort: UInt16 = 4985
+        let wsPort: UInt16 = 4984
+        let auth = ListenerPasswordAuthenticator { self.isValidCredentials($0, password: $1)}
+        
+        // tag::p2p-ws-api-urlendpointlistener-constructor[]
+        var config = URLEndpointListenerConfiguration.init(database: otherDB)
+        config.port = enableTLS ? wssPort : wsPort
+        config.disableTLS = !enableTLS
+        config.authenticator = auth
+        self.listener = URLEndpointListener.init(config: config) // <1>
+        // end::p2p-ws-api-urlendpointlistener-constructor[]
+    }
+    
+    func fMyActPeer() {
+        // tag::p2p-act-rep-func[]
+        
+        // tag::p2p-act-rep-initialize[]
+        guard let targetURL = URL(string: "wss://10.1.1.12:8092/otherDB") else {
+            fatalError("Invalid URL")
+        }
+        let targetEndpoint = URLEndpoint(url: targetURL)
+        var config = ReplicatorConfiguration(database: database, target: targetEndpoint) // <.>
+        
+        // end::p2p-act-rep-initialize[]
+        // tag::p2p-act-rep-config[]
+        // tag::p2p-act-rep-config-type[]
+        config.replicatorType = .pushAndPull
+        
+        // end::p2p-act-rep-config-type[]
+        // tag::autopurge-override[]
+        // set auto-purge behavior (here we override default)
+        config.enableAutoPurge = false // <.>
+        
+        // end::autopurge-override[]
+        // tag::p2p-act-rep-config-cont[]
+        // Configure Sync Mode
+        config.continuous = true
+        
+        // end::p2p-act-rep-config-cont[]
+        // tag::p2p-act-rep-config-self-cert[]
+        // Configure Server Security -- only accept self-signed certs
+        config.acceptOnlySelfSignedServerCertificate = true; // <.>
+        
+        // end::p2p-act-rep-config-self-cert[]
+        // Configure Client Security // <.>
+        // tag::p2p-act-rep-auth[]
+        //  Set Authentication Mode
+        config.authenticator = BasicAuthenticator(username: "cbl-user-01",
+                                                  password: "secret")
+        
+        // end::p2p-act-rep-auth[]
+        // tag::p2p-act-rep-config-conflict[]
+        /* Optionally set custom conflict resolver call back
+         config.conflictResolver = LocalWinConflictResolver()  // <.>
+         */
+        
+        // end::p2p-act-rep-config-conflict[]
+        // end::p2p-act-rep-config[]
+        // tag::p2p-act-rep-start-full[]
+        // Apply configuration settings to the replicator
+        self.replicator = Replicator.init( config: config) // <.>
+        
+        // tag::p2p-act-rep-add-change-listener[]
+        // Optionally add a change listener
+        // Retain token for use in deletion
+        let token = self.replicator.addChangeListener { change in // <.>
+            if change.status.activity == .stopped {
+                print("Replication stopped")
+            } else {
+                // tag::p2p-act-rep-status[]
+                print("Replicator is currently : \(self.replicator.status.activity)")
+            }
+        }
+        // end::p2p-act-rep-status[]
+        // end::p2p-act-rep-add-change-listener[]
+        
+        // tag::p2p-act-rep-start[]
+        // Run the replicator using the config settings
+        self.replicator.start()  // <.>
+        
+        // end::p2p-act-rep-start[]
+        // end::p2p-act-rep-start-full[]
+        
+        
+        // end::p2p-act-rep-func[]
+        self.replicator.removeChangeListener(withToken: token)
+    }
+    
+    
+    func dontTestReplicatorStop() {
+        let token = self.replicator.addChangeListener { change in }
+        // tag::p2p-act-rep-stop[]
+        
+        // Remove the change listener
+        self.replicator.removeChangeListener(withToken: token)
+        
+        // Stop the replicator
+        self.replicator.stop()
+        
+        // end::p2p-act-rep-stop[]
+    }
+    
+    func dontTestAdditionalListenerConfigs() throws {
+        let target = DatabaseEndpoint(database: otherDB)
+        var config = ReplicatorConfiguration(database: database, target: target)
+        let cert = self.listener.tlsIdentity!.certs[0]
+        let validUsername = "cbl-user-01"
+        let validPassword = "secret"
+        // tag::p2p-act-rep-config-tls-full[]
+        // tag::p2p-act-rep-config-cacert[]
+        // Configure Server Security -- only accept CA Certs
+        config.acceptOnlySelfSignedServerCertificate = false // <.>
+        
+        // end::p2p-act-rep-config-cacert[]
+        // tag::p2p-act-rep-config-self-cert[]
+        // Configure Server Security -- only accept self-signed certs
+        config.acceptOnlySelfSignedServerCertificate = true // <.>
+        
+        // end::p2p-act-rep-config-self-cert[]
+        // tag::p2p-act-rep-config-pinnedcert[]
+        // Return the remote pinned cert (the listener's cert)
+        config.pinnedServerCertificate = cert // Get listener cert if pinned
+        
+        // end::p2p-act-rep-config-pinnedcert[]
+        // Configure Client Security // <.>
+        // tag::p2p-act-rep-auth[]
+        //  Set Authentication Mode
+        config.authenticator = BasicAuthenticator(username: validUsername, password: validPassword)
+        
+        // end::p2p-act-rep-auth[]
+        // end::p2p-act-rep-config-tls-full[]
+        
+        // tag::p2p-tlsid-tlsidentity-with-label[]
+        // Check if Id exists in keychain and if so, use that Id
+        if let tlsIdentity = try TLSIdentity.identity(withLabel: "doco-sync-server") { // <.>
+            print("An identity with label : doco-sync-server already exists in keychain")
+            config.authenticator = ClientCertificateAuthenticator(identity: tlsIdentity) // <.>
+        }
+        
+        // end::p2p-tlsid-check-keychain[]
+        // end::p2p-tlsid-tlsidentity-with-label[]
+    }
+    
+    // tag::p2p-tlsid-manage-func[]
+    func myGetCert() throws -> TLSIdentity? {
+        var osStatus: OSStatus
+        let thisLabel : String? = "doco-sync-server"
+        
+        //var thisData : CFData?
+        // tag::old-p2p-tlsid-tlsidentity-with-label[]
+        // tag::p2p-tlsid-check-keychain[]
+        // USE KEYCHAIN IDENTITY IF EXISTS
+        // Check if Id exists in keychain. If so use that Id
+        if let tlsIdentity = try TLSIdentity.identity(withLabel: "doco-sync-server") {
+            print("An identity with label : doco-sync-server already exists in keychain")
+            return tlsIdentity
+        }
+        
+        // end::p2p-tlsid-check-keychain[]
+        //        thisAuthenticator.ClientCertificateAuthenticator(identity: thisIdentity )
+        //        config.thisAuthenticator
+        //        FIXME: Not sure, whats done here? Is this client side / server side authenticator?
+        // end::old-p2p-tlsid-tlsidentity-with-label[]
+        
+        
+        // tag::p2p-tlsid-check-bundled[]
+        // CREATE IDENTITY FROM BUNDLED RESOURCE IF FOUND
+        
+        // Check for a resource bundle with required label to generate identity from
+        // return nil identify if not found
+        guard let path = Bundle.main.path(forResource: "doco-sync-server", ofType: "p12"),
+              let certData = NSData(contentsOfFile: path)
+        else {
+            return nil
+        }
+        // end::p2p-tlsid-check-bundled[]
+        
+        // tag::p2p-tlsid-import-from-bundled[]
+        // Use SecPKCS12Import to import the contents (identities and certificates)
+        // of the required resource bundle (PKCS #12 formatted blob).
+        //
+        // Set passphrase using kSecImportExportPassphrase.
+        // This passphrase should correspond to what was specified when .p12 file was created
+        var result : CFArray?
+        osStatus = SecPKCS12Import(certData as CFData, [String(kSecImportExportPassphrase): "couchbase"] as CFDictionary, &result)
+        if osStatus != errSecSuccess {
+            print("Failed to import data from provided with error :\(osStatus) ")
+            return nil
+        }
+        let importedItems = result! as NSArray
+        let item = importedItems[0] as! [String: Any]
+        
+        // Get SecIdentityRef representing the item's id
+        let secIdentity = item[String(kSecImportItemIdentity)]  as! SecIdentity
+        
+        // Get Id's Private Key, return nil id if fails
+        var privateKey : SecKey?
+        osStatus = SecIdentityCopyPrivateKey(secIdentity, &privateKey)
+        if osStatus != errSecSuccess {
+            print("Failed to import private key from provided with error :\(osStatus) ")
+            return nil
+        }
+        
+        // Get all relevant certs [SecCertificate] from the ID's cert chain using kSecImportItemCertChain
+        let certChain = item[String(kSecImportItemCertChain)] as? [SecCertificate]
+        
+        
+        // Return nil, if errors in key, certChain at this stage
+        guard let privateKey = privateKey, let certChain = certChain else {
+            return nil
+        }
+        // end::p2p-tlsid-import-from-bundled[]
+        
+        // tag::p2p-tlsid-store-in-keychain[]
+        // STORE THE IDENTITY AND ITS CERT CHAIN IN THE KEYCHAIN
+#if os(iOS)
+        // For iOS, need to save the identity into the KeyChain.
+        // Save or Update identity with a label so that it could be cleaned up easily
+        // Store Private Key in Keychain
+        let params: [String : Any] = [
+            String(kSecClass):          kSecClassKey,
+            String(kSecAttrKeyType):    kSecAttrKeyTypeRSA,
+            String(kSecAttrKeyClass):   kSecAttrKeyClassPrivate,
+            String(kSecValueRef):       privateKey
+        ]
+        osStatus = SecItemAdd(params as CFDictionary, nil)
+        if osStatus != errSecSuccess {
+            print("Unable to store private key")
+            return nil
+        }
+        // Store all Certs for Id in Keychain:
+        var i = 0;
+        for cert in certChain {
+            let params: [String : Any] = [
+                String(kSecClass):      kSecClassCertificate,
+                String(kSecValueRef):   cert,
+                String(kSecAttrLabel):  "doco-sync-server"
+            ]
+            osStatus = SecItemAdd(params as CFDictionary, nil)
+            if osStatus != errSecSuccess {
+                print("Unable to store certs")
+                return nil
+            }
+            i=i+1
+        }
+#else
+        let query: [String : Any] = [
+            String(kSecClass):          kSecClassCertificate,
+            String(kSecValueRef):       certs[0]
+        ]
+        
+        let update: [String: Any] = [
+            String(kSecClass):          kSecClassCertificate,
+            String(kSecValueRef):       certs[0],
+            String(kSecAttrLabel):      label
+        ]
+        
+        osStatus = SecItemUpdate(query as CFDictionary, update as CFDictionary)
+        if osStatus != errSecSuccess {
+            print("Unable to update certs \(osStatus)")
+            return nil
+        }
+#endif
+        // end::p2p-tlsid-store-in-keychain[]
+        
+        // tag::p2p-tlsid-return-id-from-keychain[]
+        
+        // RETURN A TLSIDENTITY FROM THE KEYCHAIN FOR USE IN CONFIGURING TLS COMMUNICATION
+        return try TLSIdentity.identity(withIdentity: secIdentity, certs: [certChain[1]])
+        // end::p2p-tlsid-return-id-from-keychain[]
+    }
+    
+    func dontTestDeleteIDFromKeychain() throws {
+        // tag::p2p-tlsid-delete-id-from-keychain[]
+        
+        try TLSIdentity.deleteIdentity(withLabel: "doco-sync-server")
+        
+        // end::p2p-tlsid-delete-id-from-keychain[]
+        
+    }
+    
+    func dontTestReplicatorConfigSelfCert() throws {
+        let target = DatabaseEndpoint(database: otherDB)
+        var config = ReplicatorConfiguration(database: database, target: target)
+        // tag::old-p2p-act-rep-config-self-cert[]
+        // acceptOnlySelfSignedServerCertificate = true -- accept Slf-Signed Certs
+        //        FIXME: !!
+        //          `disableTLS` is for listener and it is 'false' by default.
+        //          `acceptOnlySelfSignedServerCertificate` is for replicator
+        //        config.disableTLS = false
+        config.acceptOnlySelfSignedServerCertificate = true
+        
+        // end::old-p2p-act-rep-config-self-cert[]
+    }
+    
+    // tag::p2p-act-rep-config-cacert-pinned-func[]
+    func myCaCertPinned() {
+        let targetURL = URL(string: "wss://10.1.1.12:8092/otherDB")!
+        let targetEndpoint = URLEndpoint(url: targetURL)
+        var config = ReplicatorConfiguration(database: database, target: targetEndpoint)
+        // tag::p2p-act-rep-config-cacert-pinned[]
+        
+        // Get bundled resource and read into localcert
+        guard
+            let pathToCert = Bundle.main.path(forResource: "listener-pinned-cert", ofType: "cer"),
+            let localCertificate:NSData = NSData(contentsOfFile: pathToCert)
+        else { /* process error */ return }
+        
+        // Create certificate
+        // using its DER representation as a CFData
+        guard
+            let pinnedCert = SecCertificateCreateWithData(nil, localCertificate)
+        else { /* process error */  return }
+        
+        // Add `pinnedCert` and `acceptOnlySelfSignedServerCertificate=false` to `ReplicatorConfiguration`
+        config.acceptOnlySelfSignedServerCertificate = false
+        config.pinnedServerCertificate = pinnedCert
+        // end::p2p-act-rep-config-cacert-pinned[]
+        // end::p2p-act-rep-config-cacert-pinned-func[]
+        
+    }
+    
+    func dontTestOldListenerConfigClientRootCA() throws {
+        // tag::old-listener-config-client-root-ca[]
+        // Configure the client authenticator to validate using ROOT CA <.>
+        // end::old-listener-config-client-root-ca[]
+    }
+    
+    // end::p2p-tlsid-manage-func[]
+    
+    // tag::replication-start-func[]
+    enum PeerConnectionStatus: UInt8 {
+        case stopped = 0;
+        case offline
+        case connecting
+        case idle
+        case busy
+    }
+    
+    func dontTestReplicationStart(_ peer: String,
+                                  peerDBName: String,
+                                  user: String?,
+                                  pass: String?,
+                                  handler: @escaping (PeerConnectionStatus, Error?) -> Void) throws {
+        guard let userDb = self.database else {
+            fatalError("DatabaseNotInitialized")
+            // ... take appropriate actions
+        }
+        guard let validUser = user, let validPassword = pass else {
+            fatalError("UserCredentialsNotProvided")
+            // ... take appropriate actions
+        }
+        
+        // tag::replicator-start-func-config-init[]
+        let replicator = self.replicatorsToPeers[peer]
+        
+        if replicator == nil {
+            // Start replicator to connect to the URLListenerEndpoint
+            guard let targetUrl = URL(string: "ws://\(peer)/\(peerDBName)") else {
+                fatalError("URLInvalid")
+                // ... take appropriate actions
+            }
+            
+            var config = ReplicatorConfiguration.init(database: userDb, target: URLEndpoint.init(url:targetUrl)) //<1>
+            // end::replicator-start-func-config-init[]
+            
+            // tag::replicator-start-func-config-more[]
+            
+            config.replicatorType = .pushAndPull // <2>
+            config.continuous =  true // <3>
+            
+            // end::replicator-start-func-config-more[]
+            
+            // tag::replicator-start-func-config-auth[]
+            config.authenticator = BasicAuthenticator(username: validUser, password: validPassword)
+            // end::replicator-start-func-config-auth[]
+            
+            // tag::replicator-start-func-repl-init[]
+            let temp = Replicator.init(config: config)
+            self.replicatorsToPeers[peer] = temp
+            
+            let token = registerForEventsForReplicator(temp, handler: handler)
+            self.replicatorListenerTokens[peer] = token
+            
+            // end::replicator-start-func-repl-init[]
+        }
+        
+        // tag::replicator-start-func-repl-start[]
+        replicator?.start()
+        // end::replicator-start-func-repl-start[]
+    }
+    // end::replication-start-func[]
+    
+    // tag::replicator-register-for-events[]
+    func registerForEventsForReplicator(_ replicator: Replicator,
+                                        handler: @escaping (PeerConnectionStatus, Error?) -> Void) -> ListenerToken {
+        return replicator.addChangeListener { change in
+            guard change.status.error == nil else {
+                handler(.stopped, change.status.error)
+                return
+            }
+            
+            switch change.status.activity {
+            case .connecting:
+                print("Replicator Connecting to Peer")
+            case .idle:
+                print("Replicator in Idle state")
+            case .busy:
+                print("Replicator in busy state")
+            case .offline:
+                print("Replicator in offline state")
+            case .stopped:
+                print("Replicator is stopped")
+            }
+            
+            let progress = change.status.progress
+            if progress.completed == progress.total {
+                print("All documents synced")
+            }
+            else {
+                print("Documents \(progress.total - progress.completed) still pending sync")
+            }
+            
+            if let customStatus = PeerConnectionStatus(rawValue: change.status.activity.rawValue) {
+                handler(customStatus, nil)
+            }
+        }
+    }
+    // end::replicator-register-for-events[]
 }
 
 // tag::sgw-repl-pull[]
 class MyClass {
     var database: Database!
     var replicator: Replicator! // <1>
-
+    
     func startReplicator() {
         let url = URL(string: "ws://localhost:4984/db")! // <2>
         let target = URLEndpoint(url: url)
         var config = ReplicatorConfiguration(database: database, target: target)
         config.replicatorType = .pull
-
+        
         self.replicator = Replicator(config: config)
         self.replicator.start()
     }
@@ -59,22 +2466,22 @@ class MyClass {
 
 /*
  // tag::sgw-repl-pull-callouts[]
-
+ 
  <.> A replication is an asynchronous operation.
  To keep a reference to the `replicator` object, you can set it as an instance property.
  <.> The URL scheme for remote database URLs has changed in Couchbase Lite 2.0.
  You should now use `ws:`, or `wss:` for SSL/TLS connections.
-
-
+ 
+ 
  // end::sgw-repl-pull-callouts[]
-
+ 
  // tag::sgw-act-rep-initialize[]
  let tgtUrl = URL(string: "wss://10.1.1.12:8092/travel-sample")!
  let targetEndpoint = URLEndpoint(url: tgtUrl)
  var config = ReplicatorConfiguration(database: database!, target: targetEndpoint) // <.>
-
+ 
  // end::sgw-act-rep-initialize[]
-
+ 
  */
 
 // MARK: -- Conflict Resolver Helpers
@@ -147,19 +2554,19 @@ class TestPredictiveModel: PredictiveModel {
         numberOfCalls = numberOfCalls + 1
         return self.doPredict(input: input)
     }
-
+    
     func doPredict(input: DictionaryObject) -> DictionaryObject? {
         return nil
     }
-
+    
     func registerModel() {
         Database.prediction.registerModel(self, withName: type(of: self).name)
     }
-
+    
     func unregisterModel() {
         Database.prediction.unregisterModel(withName: type(of: self).name)
     }
-
+    
     func reset() {
         numberOfCalls = 0
     }
@@ -185,6 +2592,199 @@ class LogTestLogger: Logger {
 }
 // end::custom-logging[]
 
+struct Hotel: Codable {
+    var id: String
+    var type: String?
+    var name: String?
+    var city: String?
+}
+
+#if os(macOS)
+// tag::listener-get-network-interfaces[]
+import SystemConfiguration
+// . . .
+
+class SomeClass {
+    func SomeFunction() {
+        for interface in SCNetworkInterfaceCopyAll() as! [SCNetworkInterface] {
+            // do something with this `interface`
+        }
+    }
+    
+    // . . .
+}
+
+// end::listener-get-network-interfaces[]
+#endif
+
+// MARK -- P2p
+
+/* ----------------------------------------------------------- */
+/* ---------------------  ACTIVE SIDE  ----------------------- */
+/* ---------------  stubs for documentation  ----------------- */
+/* ----------------------------------------------------------- */
+class ActivePeer: MessageEndpointDelegate {
+    
+    init() throws {
+        let id = ""
+        
+        // tag::message-endpoint[]
+        let database = try Database(name: "dbname")
+        
+        // The delegate must implement the `MessageEndpointDelegate` protocol.
+        let messageEndpointTarget = MessageEndpoint(uid: "UID:123", target: id, protocolType: .messageStream, delegate: self)
+        // end::message-endpoint[]
+        
+        // tag::message-endpoint-replicator[]
+        let config = ReplicatorConfiguration(database: database, target: messageEndpointTarget)
+        
+        // Create the replicator object.
+        let replicator = Replicator(config: config)
+        // Start the replication.
+        replicator.start()
+        // end::message-endpoint-replicator[]
+    }
+    
+    // tag::create-connection[]
+    /* implementation of MessageEndpointDelegate */
+    func createConnection(endpoint: MessageEndpoint) -> MessageEndpointConnection {
+        let connection = ActivePeerConnection() /* implements MessageEndpointConnection */
+        return connection
+    }
+    // end::create-connection[]
+    
+}
+
+class ActivePeerConnection: MessageEndpointConnection {
+    
+    var replicatorConnection: ReplicatorConnection?
+    
+    init() {}
+    
+    func disconnect() {
+        // tag::active-replicator-close[]
+        replicatorConnection?.close(error: nil)
+        // end::active-replicator-close[]
+    }
+    
+    // tag::active-peer-open[]
+    /* implementation of MessageEndpointConnection */
+    func open(connection: ReplicatorConnection, completion: @escaping (Bool, MessagingError?) -> Void) {
+        replicatorConnection = connection
+        completion(true, nil)
+    }
+    // end::active-peer-open[]
+    
+    // tag::active-peer-send[]
+    /* implementation of MessageEndpointConnection */
+    func send(message: Message, completion: @escaping (Bool, MessagingError?) -> Void) {
+        var data = message.toData()
+        /* send the data to the other peer */
+        /* ... */
+        /* call the completion handler once the message is sent */
+        completion(true, nil)
+    }
+    // end::active-peer-send[]
+    
+    func receive(data: Data) {
+        // tag::active-peer-receive[]
+        let message = Message.fromData(data)
+        replicatorConnection?.receive(message: message)
+        // end::active-peer-receive[]
+    }
+    
+    // tag::active-peer-close[]
+    /* implementation of MessageEndpointConnection */
+    func close(error: Error?, completion: @escaping () -> Void) {
+        /* disconnect with communications framework */
+        /* ... */
+        /* call completion handler */
+        completion()
+    }
+    // end::active-peer-close[]
+    
+}
+
+/* ----------------------------------------------------------- */
+/* ---------------------  PASSIVE SIDE  ---------------------- */
+/* ---------------  stubs for documentation  ----------------- */
+/* ----------------------------------------------------------- */
+class PassivePeerConnection: NSObject, MessageEndpointConnection {
+    
+    var messageEndpointListener: MessageEndpointListener?
+    var replicatorConnection: ReplicatorConnection?
+    
+    override init() {
+        super.init()
+    }
+    
+    func startListener() {
+        // tag::listener[]
+        let database = try! Database(name: "mydb")
+        let config = MessageEndpointListenerConfiguration(database: database, protocolType: .messageStream)
+        messageEndpointListener = MessageEndpointListener(config: config)
+        // end::listener[]
+    }
+    
+    func stopListener() {
+        // tag::passive-stop-listener[]
+        messageEndpointListener?.closeAll()
+        // end::passive-stop-listener[]
+    }
+    
+    func acceptConnection() {
+        // tag::advertizer-accept[]
+        let connection = PassivePeerConnection() /* implements MessageEndpointConnection */
+        messageEndpointListener?.accept(connection: connection)
+        // end::advertizer-accept[]
+    }
+    
+    func disconnect() {
+        // tag::passive-replicator-close[]
+        replicatorConnection?.close(error: nil)
+        // end::passive-replicator-close[]
+    }
+    
+    // tag::passive-peer-open[]
+    /* implementation of MessageEndpointConnection */
+    func open(connection: ReplicatorConnection, completion: @escaping (Bool, MessagingError?) -> Void) {
+        replicatorConnection = connection
+        completion(true, nil)
+    }
+    // end::passive-peer-open[]
+    
+    // tag::passive-peer-send[]
+    /* implementation of MessageEndpointConnection */
+    func send(message: Message, completion: @escaping (Bool, MessagingError?) -> Void) {
+        var data = Data()
+        data.append(message.toData())
+        /* send the data to the other peer */
+        /* ... */
+        /* call the completion handler once the message is sent */
+        completion(true, nil)
+    }
+    // end::passive-peer-send[]
+    
+    func receive(data: Data) {
+        // tag::passive-peer-receive[]
+        let message = Message.fromData(data)
+        replicatorConnection?.receive(message: message)
+        // end::passive-peer-receive[]
+    }
+    
+    // tag::passive-peer-close[]
+    /* implementation of MessageEndpointConnection */
+    func close(error: Error?, completion: @escaping () -> Void) {
+        /* disconnect with communications framework */
+        /* ... */
+        /* call completion handler */
+        completion()
+    }
+    // end::passive-peer-close[]
+}
+
+
+
 // MARK: -- QUESTIONS
 
 // tag::listener[]
@@ -201,3 +2801,10 @@ class LogTestLogger: Logger {
 // https://docs.couchbase.com/mobile/2.8.0/couchbase-lite-swift/Classes/URLEndpointListener.html
 // end::p2p-ws-api-urlendpointlistener[]
 
+// tag::1x-attachment[]
+// FIXME: Removed!!
+// end::1x-attachment[]
+
+// tag::tojson-getblobasstring[]
+// FIXME: this tag is same as tag[tojson-blob[]]
+// end::tojson-getblobasstring[]
